chore(#SFD2-475): add environment for Comms specific scenario

+++ entrypoint.sh
@@ -17,12 +17,12 @@ mkdir -p ${JM_REPORTS} ${JM_LOGS}
 
 # Clean up reports directory if it exists and is not empty
 if [ -d "${JM_REPORTS}" ] && [ "$(ls -A ${JM_REPORTS})" ]; then
-  rm -rf ${JM_REPORTS}/*
+  rm -rf "${JM_REPORTS:?}/*"
 fi
 
 TEST_SCENARIO=${TEST_SCENARIO:-test}
 SCENARIOFILE=${JM_SCENARIOS}/${TEST_SCENARIO}.jmx
-REPORTFILE=${NOW}-perftest-${TEST_SCENARIO}-report.csv
+REPORTFILE=${NOW}-perftest-${TEST_SCENARIO}-report.jtl
 LOGFILE=${JM_LOGS}/perftest-${TEST_SCENARIO}.log
 
 # Before running the suite, replace 'service-name' with the name/url of the service to test.
@@ -31,14 +31,29 @@ SERVICE_ENDPOINT=${SERVICE_ENDPOINT:-service-name.${ENVIRONMENT}.cdp-int.defra.c
 # PORT is used to set the port of this performance test container
 SERVICE_PORT=${SERVICE_PORT:-443}
 SERVICE_URL_SCHEME=${SERVICE_URL_SCHEME:-https}
+LOOP_COUNT=${LOOP_COUNT:-100}
+DURATION_SECONDS=${DURATION_SECONDS:-300}
+THREAD_COUNT=${THREAD_COUNT:-1}
+RAMPUP_SECONDS=${RAMPUP_SECONDS:-30}
 
 # Run the test suite
 jmeter -n -t ${SCENARIOFILE} -e -l "${REPORTFILE}" -o ${JM_REPORTS} -j ${LOGFILE} \
 -Jenv="${ENVIRONMENT}" \
 -Jdomain="${SERVICE_ENDPOINT}" \
 -Jport="${SERVICE_PORT}" \
--Jprotocol="${SERVICE_URL_SCHEME}"
+-Jprotocol="${SERVICE_URL_SCHEME}" \
+-JRAMPUP_SECONDS="${RAMPUP_SECONDS}" \
+-JTHREAD_COUNT="${THREAD_COUNT}" \
+-JDURATION_SECONDS="${DURATION_SECONDS}" \
+-JLOOP_COUNT="${LOOP_COUNT}"
 
+
+# list content of reports directory and echo
+search_dir="${JM_REPORTS}"
+for entry in "$search_dir"/*
+do
+  echo "$entry"
+done
 # Publish the results into S3 so they can be displayed in the CDP Portal
 if [ -n "$RESULTS_OUTPUT_S3_PATH" ]; then
   # Copy the CSV report file and the generated report files to the S3 bucket
@@ -57,4 +72,5 @@ else
    exit 1
 fi
 
-exit $test_exit_code
+echo "Test execution completed with exit code $?"
+exit 0
